Added caveat to LTCE page - translation updates - static code/data
